app/backend/requirements.txt
absl-py==2.1.0
astunparse==1.6.3
audeer==2.2.0
audformat==1.3.1
audinterface==1.2.2
audiofile==1.5.0
audioread==3.0.1
audmath==1.4.1
audobject==0.7.11
audresample==1.3.3
blinker==1.8.2
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.4.0
click==8.1.7
decorator==5.1.1
filelock==3.16.1
Flask==3.0.3
Flask-Cors==5.0.0
flatbuffers==24.3.25
fsspec==2024.10.0
gast==0.6.0
google-pasta==0.2.0
grpcio==1.67.1
h5py==3.12.1
huggingface-hub==0.26.2
idna==3.10
importlib_metadata==8.5.0
iso3166==2.1.1
iso639-lang==2.5.1
itsdangerous==2.2.0
Jinja2==3.1.4
joblib==1.4.2
keras==3.6.0
lazy_loader==0.4
libclang==18.1.1
librosa==0.10.2.post1
llvmlite==0.43.0
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
ml-dtypes==0.3.2
more-itertools==10.5.0
mpmath==1.3.0
msgpack==1.1.0
namex==0.0.8
networkx==3.4.2
numba==0.60.0
numpy==1.26.4
openai-whisper==20240930
opensmile==2.5.0
opt_einsum==3.4.0
optree==0.13.0
oyaml==1.0
packaging==24.1
pandas==2.2.3
platformdirs==4.3.6
pooch==1.8.2
protobuf==4.25.5
pyarrow==18.1.0
pycparser==2.22
pydub==0.25.1
Pygments==2.18.0
python-dateutil==2.9.0.post0
pytz==2024.2
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.3
rich==13.9.4
safetensors==0.4.5
scikit-learn==1.5.2
scipy==1.14.1
setuptools==75.3.0
six==1.16.0
sounddevice==0.5.1
soundfile==0.12.1
soxr==0.5.0.post1
sympy==1.13.3
tensorboard==2.16.2
tensorboard-data-server==0.7.2
tensorflow==2.16.2
termcolor==2.5.0
threadpoolctl==3.5.0
tiktoken==0.8.0
tokenizers==0.20.3
torch==2.2.2
tqdm==4.67.0
transformers==4.46.2
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3
Werkzeug==3.1.2
wheel==0.44.0
wrapt==1.16.0
zipp==3.21.0
